move SeatSelection in TicketSelection
