+++ src/components/EventDetail/TicketSelection/SeatSelection/PromotionList.tsx
